Add: Mobil Layout Switch
Added logic for mobil content layout.
Uses the same as in detect_ie(), but
is independent. Can be turned on/off in Backend.

Its stored in its own function detect_mobil()
and returns true if the user is using a smartphone.

We won't have cookies here, since there is already
a browser implementation for switching mobil/desktop.