Prepare Compatibility for Joomla 1.5 (Part 1)
The Template should run on Joomla 1.5 as well as on 2.5(+).
This patch series will introduce compatibility for Joomla 1.5.

J1.5 stores the backend information in a file called params.ini
instead of the database. So i added it with very basic information.

Also the languange handling is quite different in J1.5, it uses
different files for front- and backend. So this patch splits them up.

J1.5 has no use for template_preview.png and error.php (does 2.5 even
use all of our php-files?), better clean them up. Then there should be
various fixes for the Joomla core (in parameters: field becomes param).
This messes up parameter, but we will find a workaround.

What needs to be done?

* Testing!!! (i have no J1.5 enviroment)
* index.php-patch (need the J1.5 stuff)

Signed-off-by: Alexander Christ <alex.christ@hotmail.de>